ingesta/postgresql-reservations-api/requirements.txt
psycopg2-binary==2.9.7
pandas==2.1.1
boto3==1.28.62
python-dotenv==1.0.0
botocore==1.31.62